Initial Installation Filtering Issue
Upon initial installation, the filtering of search results is enabled by default. However, the filtering does not function correctly until the user manually toggles the filtering option off and then on again in the options page. Users expect the filtering to work immediately after adding or importing domains to the blacklist, without additional steps.

The extension stops filtering search results automatically at random. Users would have to go to the options page, turn the filtering switch off and then on again, and return to the tab where the search query was made for the appropriate search results to be filtered.